Modifying bidding table. Preparing Board class for playing stage.
Changing display of bidding table. Now it will adjust and rotate for specific player. Adding some useful attributes to Board class for handling playing stage. Changing value of card according to the bidded contract. Make_move method for removing card from hand and moving it on the table. Need to work for comparision of cards in trick (if higher card is not suited to the lead suit).